ingestion/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1

# Document processing
pypdf>=3.17.0

# Vector database
pinecone-client>=3.0.0

# Environment management
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0

